+++ Solutions/PA1/test/check.cpp
@@ -98,7 +98,7 @@ int main(void) {
   double grade = num_correct / 10;
   std::cout << "Grade: " << (grade * 100) << "%" << std::endl;
   if(num_correct != 10) {
-    for(int i = 0; i < results.size(); ++i) {
+    for(size_t i = 0; i < results.size(); ++i) {
       auto val = results[i];
       if(!val) {
         std::cout << "Re-run test #" << i << "." << std::endl;
